Feature: Implement error handling
Before, it was easy to make a mistake so that the number of headers didn't match the number of cells in each row. In that case, the list of CSV values results in a malformed table.

Now, there's error handling that alerts when the number of cells don't match the number of table headers.